migrate d500/depth-mapping tests to pytest
